debug

+++ client/robot_code/source/utils.c
@@ -144,6 +144,6 @@ void update_with_offset(Ray* ray) {
 float cal_run_time(int distance, int speed) {
     int count_per_rot;
     get_tacho_count_per_rot(components[LEFT_MOTOR], &count_per_rot);
-    return (float) (1000 * count_per_rot * abs(distance) )/(speed * PI * WHEEL_DIAM) ;
+    return (int) (1000 * count_per_rot * abs(distance) )/(speed * PI * WHEEL_DIAM) ;
 }
